LPS-42926 Provide a default AuthVerifier impl so that all AutoLogin's can be used in AuthVerifierPipeline
