+++ portal-impl/src/com/liferay/portal/security/auth/ParameterAutoLogin.java
@@ -24,8 +24,6 @@ import com.liferay.portal.security.pwd.PasswordEncryptorUtil;
 import com.liferay.portal.service.UserLocalServiceUtil;
 import com.liferay.portal.util.PortalUtil;
 
-import java.util.Properties;
-
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
@@ -33,37 +31,7 @@ import javax.servlet.http.HttpServletResponse;
  * @author Minhchau Dang
  * @author Tomas Polesovsky
  */
-public class ParameterAutoLogin extends BaseAutoLogin implements AuthVerifier {
-
-	@Override
-	public String getAuthType() {
-		return ParameterAutoLogin.class.getSimpleName();
-	}
-
-	@Override
-	public AuthVerifierResult verify(
-			AccessControlContext accessControlContext, Properties properties)
-		throws AuthException {
-
-		try {
-			AuthVerifierResult authVerifierResult = new AuthVerifierResult();
-
-			String[] credentials = login(
-				accessControlContext.getRequest(),
-				accessControlContext.getResponse());
-
-			if (credentials != null) {
-				authVerifierResult.setPassword(credentials[1]);
-				authVerifierResult.setState(AuthVerifierResult.State.SUCCESS);
-				authVerifierResult.setUserId(Long.valueOf(credentials[0]));
-			}
-
-			return authVerifierResult;
-		}
-		catch (AutoLoginException ale) {
-			throw new AuthException(ale);
-		}
-	}
+public class ParameterAutoLogin extends BaseAutoLogin {
 
 	@Override
 	protected String[] doLogin(
