+++ portal-service/src/com/liferay/portal/security/auth/BaseAutoLogin.java
@@ -17,13 +17,20 @@ package com.liferay.portal.security.auth;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 
+import java.util.Properties;
+
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 /**
  * @author Mate Thurzo
  */
-public abstract class BaseAutoLogin implements AutoLogin {
+public abstract class BaseAutoLogin implements AuthVerifier, AutoLogin {
+
+	@Override
+	public String getAuthType() {
+		return this.getClass().getSimpleName();
+	}
 
 	@Override
 	public String[] handleException(
@@ -44,6 +51,31 @@ public abstract class BaseAutoLogin implements AutoLogin {
 		}
 		catch (Exception e) {
 			return handleException(request, response, e);
+		}
+	}
+
+	@Override
+	public AuthVerifierResult verify(
+			AccessControlContext accessControlContext, Properties properties)
+		throws AuthException {
+
+		try {
+			AuthVerifierResult authVerifierResult = new AuthVerifierResult();
+
+			String[] credentials = login(
+				accessControlContext.getRequest(),
+				accessControlContext.getResponse());
+
+			if (credentials != null) {
+				authVerifierResult.setPassword(credentials[1]);
+				authVerifierResult.setState(AuthVerifierResult.State.SUCCESS);
+				authVerifierResult.setUserId(Long.valueOf(credentials[0]));
+			}
+
+			return authVerifierResult;
+		}
+		catch (AutoLoginException ale) {
+			throw new AuthException(ale);
 		}
 	}
